I worked on main website and CSS file
Nikki gave me feedback on the index file about how the styling should be in the CSS file not the main website, so I moved all of that over. Also, I changed the sizes of the images, and worked on the contact list at the bottom of the page. 
I hadn't made any commits lately so this one has a lot more change then most of the other file uploads.